requirements.txt
Flask==0.10.1
Flask-Script==0.6.6
Flask-WTF==0.9.4
#Flask-Security==1.6.9
Flask-SQLAlchemy==1.0
MySQL-python==1.2.5